+++ packages/server/src/lib/errors.ts
