+++ packages/server/src/middlewares/authenticate.ts
