+++ packages/server/src/models/capsule.ts
