+++ packages/server/src/schemas/image.ts
